Added Protocols and unit tested
